fs/fork.c: Modify the new task's signal, not current, on a copy
When copying a process, on a new thread the current code increments
current's signal nr_threads, live, and sigcnt values.  Currently
p->signal and current->signal are the same in the code, but with
live application dump, they will be different and the new thread's
values are really the proper ones to increment, anyway.

Signed-off-by: Corey Minyard <cminyard@mvista.com>