livedump: fix task put in do_livedump
do_livedump was expecting to do the task_put on the passed
in task, but this was breaking ptrace, and it was wrong,
anyway.

The caller should do the get and put.  So fix the proc
interface to do that, and also fix it to only request
a dump once on input.

Signed-off-by: Corey Minyard <cminyard@mvista.com>