Make down_write_killable() return -ERESTARTSYS
It used to return (or cause to be returned) -EINTR if woken.  It
didn't really matter what it returned as long as it was an error,
as the process would normally never run again.

Live application coredumps, however, use a specially handled SIGKILL
to wake the threads being dumped.  This special SIGKILL could cause
a -EINTR error to be returned from system calls that aren't supposed
to return -EINTR.

Returning -ERESTARTSYS causes the system call to be restarted as
it should be in this case.  And there is no harm in doing this
otherwise.

Signed-off-by: Corey Minyard <cminyard@mvista.com>