livedump: Fix problem handling new threads in a livedump
When cloning a thread, the livedump code was looking at the new thread,
not the current thread, to see if it was in a livedump.  Obviously the
new thread is not going to be in a livedump, so it would slip through
and possibly cause bad things to happen if it did.

Use the current thread, not the new one, to extract the dump variable.

Signed-off-by: Corey Minyard <cminyard@mvista.com>