livedump: Fix a deletion race in do_exit
When a normal (not livedump clone) thread exits, it checks to see if
it needs to wait for the clone operation to finish, and it handles
signals during this wait so it can actually process the signal to
do the clone.

There was a race where the signal handling could finish the livedump
processing and the dump variable be released.  But code would then
loop around and wait using a completion in the dump variable, resulting
in an oops.

Fix the issue by getting a ref to the dump variable during this
operation, so the dump variable isn't released.

Signed-off-by: Corey Minyard <cminyard@mvista.com>