livedump: Switch from using SIGKILL to using special sigpending
SIGKILL was just causing too many issues.  It could cause wakeups in
places that don't expect it, and blocking and unblocking is messy.

Instead, create special livedump sigpending handling that allows
livedump to have its own non-signal wakeup.

This does mean that a livedump will not wake a thread that is waiting
on a xxx_killable() function, but waking killable functions with
livedump caused some functions that are not supposed to to return -EINTR,
and caused userspace problems.

You could have all killable wakeups return -ERESTARTSYS.  That's a bigger
job, affects lots of stuff, and it's hard to tell if that's ok, if the
system calls can actually be restarted in every case.  So this change
seems the better route.

Signed-off-by: Corey Minyard <cminyard@mvista.com>