livedump: Restructure the way the code works
Rework the code so that it always goes through the main signal
handling and there is no loop waiting for signals to come in
inside the livedump code itself.  This eliminates a lot of
code and allowed a lot of simplification.

This required modifying the code to not wait in exit.  An exit
will always remove the thread from the livedump group.  But that
was also a simplification.

At the moment, a thread can no longer livedump itself.  That
should be fixable in the future.

Signed-off-by: Corey Minyard <cminyard@mvista.com>