livedump: Fix a race when a sub-thread takes a livedump
When a thread takes a livedump of itself, it signals all the threads
to stop first.  If it's the thread leader, it just continues processing.
If it's a sub-thread, it increments the number of stopping threads then
acts like it was signalled.

For the sub-thread situation, it is possible that all the other threads
run through their signal handlers and decrement the number of stopping
threads to zero before the thread increments the number of stopping
threads for itself.  This will cause the dump to continue prematurely
with an incorrect number of stopped threads, resulting in a hang.

For this specific situation, move the increment for the current thread
to before where the other threads are signalled to avoid this situation.

Also add a comment on why this can't happen when doing the signal
process.

Signed-off-by: Corey Minyard <cminyard@mvista.com>