livedump: Don't set pid_ns_for_children to override namespace
Livedump was setting pid_ns_for_children in nsproxy each time it cloned
a process.  But nsproxy is common for all threads in a process, so these
would step on each other and leave the final value for each thread in a
possible bad state.

So instead, pass in the namespace to the kernel clone arguments instead.
Also pass in the internal flags instead of passing them on the
copy_process() command line, just to keep things neater.

Signed-off-by: Corey Minyard <cminyard@mvista.com>